Refactored TlsFixture to reduce duplication
