+++ ledger/participant-integration-api/BUILD.bazel
@@ -159,7 +159,9 @@ da_scala_library(
         "//daml-lf/transaction-test-lib",
         "//language-support/scala/bindings",
         "//ledger-api/rs-grpc-bridge",
+        "//ledger-api/sample-service",
         "//ledger-api/testing-utils",
+        "//ledger/ledger-api-client",
         "//ledger/ledger-api-common",
         "//ledger/ledger-api-domain",
         "//ledger/ledger-api-health",
