+++ ledger/participant-integration-api/src/test/suite/scala/platform/apiserver/tls/TlsCertificateRevocationCheckingSpec.scala
@@ -4,32 +4,14 @@
 package com.daml.platform.apiserver.tls
 
 import java.io.File
-import java.util.concurrent.Executors
 
-import com.codahale.metrics.MetricRegistry
 import com.daml.bazeltools.BazelRunfiles.rlocation
-import com.daml.grpc.sampleservice.implementations.ReferenceImplementation
 import com.daml.ledger.api.testing.utils.AkkaBeforeAndAfterAll
-import com.daml.ledger.api.tls.TlsConfiguration
-import com.daml.ledger.client.GrpcChannel
-import com.daml.ledger.client.configuration.{
-  CommandClientConfiguration,
-  LedgerClientConfiguration,
-  LedgerIdRequirement,
-}
-import com.daml.ledger.resources.{Resource, ResourceContext, ResourceOwner, TestResourceContext}
-import com.daml.logging.LoggingContext
-import com.daml.metrics.Metrics
-import com.daml.platform.apiserver.{ApiServer, ApiServices, LedgerApiServer}
-import com.daml.platform.hello.{HelloRequest, HelloResponse, HelloServiceGrpc}
-import com.daml.ports.Port
-import io.grpc.{BindableService, ManagedChannel}
+import com.daml.ledger.resources.TestResourceContext
+import com.daml.platform.apiserver.LedgerApiServer
 import org.mockito.MockitoSugar
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AsyncWordSpec
-
-import scala.collection.immutable
-import scala.concurrent.Future
 
 final class TlsCertificateRevocationCheckingSpec
     extends AsyncWordSpec
@@ -38,7 +20,7 @@ final class TlsCertificateRevocationCheckingSpec
     with AkkaBeforeAndAfterAll
     with TestResourceContext
     with OcspResponderFixture {
-  import TlsCertificateRevocationCheckingSpec.{TlsFixture, resource}
+  import TlsCertificateRevocationCheckingSpec.resource
 
   val serverCrt = resource("server.crt")
   val serverKey = resource("server.pem")
@@ -60,7 +42,15 @@ final class TlsCertificateRevocationCheckingSpec
   classOf[LedgerApiServer].getSimpleName when {
     "certificate revocation checking is enabled" should {
       "allow TLS connections with valid certificates" in {
-        TlsFixture(tlsEnabled = true, serverCrt, serverKey, caCrt, clientCrt, clientKey)
+        TlsFixture(
+          tlsEnabled = true,
+          serverCrt,
+          serverKey,
+          caCrt,
+          Some(clientCrt),
+          Some(clientKey),
+          certRevocationChecking = true,
+        )
           .makeARequest()
           .map(_ => succeed)
       }
@@ -71,8 +61,9 @@ final class TlsCertificateRevocationCheckingSpec
           serverCrt,
           serverKey,
           caCrt,
-          clientRevokedCrt,
-          clientRevokedKey,
+          Some(clientRevokedCrt),
+          Some(clientRevokedKey),
+          certRevocationChecking = true,
         )
           .makeARequest()
           .failed
@@ -87,7 +78,14 @@ final class TlsCertificateRevocationCheckingSpec
 
     "certificate revocation checking is not enabled" should {
       "allow TLS connections with valid certificates" in {
-        TlsFixture(tlsEnabled = false, serverCrt, serverKey, caCrt, clientCrt, clientKey)
+        TlsFixture(
+          tlsEnabled = false,
+          serverCrt,
+          serverKey,
+          caCrt,
+          Some(clientCrt),
+          Some(clientKey),
+        )
           .makeARequest()
           .map(_ => succeed)
       }
@@ -98,8 +96,8 @@ final class TlsCertificateRevocationCheckingSpec
           serverCrt,
           serverKey,
           caCrt,
-          clientRevokedCrt,
-          clientRevokedKey,
+          Some(clientRevokedCrt),
+          Some(clientRevokedKey),
         )
           .makeARequest()
           .map(_ => succeed)
@@ -110,90 +108,7 @@ final class TlsCertificateRevocationCheckingSpec
 
 object TlsCertificateRevocationCheckingSpec {
 
-  protected final case class TlsFixture(
-      tlsEnabled: Boolean,
-      serverCrt: File,
-      serverKey: File,
-      caCrt: File,
-      clientCrt: File,
-      clientKey: File,
-  )(implicit rc: ResourceContext) {
-
-    def makeARequest(): Future[HelloResponse] =
-      resources().use { channel =>
-        val testRequest = HelloRequest(1)
-        HelloServiceGrpc
-          .stub(channel)
-          .single(testRequest)
-      }
-
-    private val DefaultMaxInboundMessageSize: Int = 4 * 1024 * 1024 // taken from the Sandbox config
-
-    private final class MockApiServices(apiServices: ApiServices)
-        extends ResourceOwner[ApiServices] {
-      override def acquire()(implicit context: ResourceContext): Resource[ApiServices] = {
-        Resource(Future.successful(apiServices))(_ => Future.successful(()))(context)
-      }
-    }
-
-    private final class EmptyApiServices extends ApiServices {
-      override val services: Iterable[BindableService] = List(new ReferenceImplementation)
-      override def withServices(otherServices: immutable.Seq[BindableService]): ApiServices = this
-    }
-
-    private val serverTlsConfiguration = TlsConfiguration(
-      enabled = tlsEnabled,
-      keyCertChainFile = Some(serverCrt),
-      keyFile = Some(serverKey),
-      trustCertCollectionFile = Some(caCrt),
-      enableCertRevocationChecking = true,
-    )
-
-    private def apiServerOwner(): ResourceOwner[ApiServer] = {
-      val apiServices = new EmptyApiServices
-      val owner = new MockApiServices(apiServices)
-
-      LoggingContext.newLoggingContext { implicit loggingContext =>
-        ResourceOwner
-          .forExecutorService(() => Executors.newCachedThreadPool())
-          .flatMap(servicesExecutor =>
-            new LedgerApiServer(
-              apiServicesOwner = owner,
-              desiredPort = Port.Dynamic,
-              maxInboundMessageSize = DefaultMaxInboundMessageSize,
-              address = None,
-              tlsConfiguration = Some(serverTlsConfiguration),
-              servicesExecutor = servicesExecutor,
-              metrics = new Metrics(new MetricRegistry),
-            )
-          )
-      }
-    }
-
-    private val clientTlsConfiguration =
-      TlsConfiguration(
-        enabled = tlsEnabled,
-        keyCertChainFile = Some(clientCrt),
-        keyFile = Some(clientKey),
-        trustCertCollectionFile = Some(caCrt),
-      )
-
-    private val ledgerClientConfiguration = LedgerClientConfiguration(
-      applicationId = s"TlsCertificates-app",
-      ledgerIdRequirement = LedgerIdRequirement.none,
-      commandClient = CommandClientConfiguration.default,
-      sslContext = clientTlsConfiguration.client,
-      token = None,
-    )
-
-    private def resources(): ResourceOwner[ManagedChannel] =
-      for {
-        apiServer <- apiServerOwner()
-        channel <- new GrpcChannel.Owner(apiServer.port, ledgerClientConfiguration)
-      } yield channel
-
-  }
-
   protected def resource(src: String) =
     new File(rlocation("ledger/test-common/test-certificates/" + src))
+
 }
